19-03-2022

+++ lib/services/sell_asset.dart
@@ -15,24 +15,28 @@ Future<String> sellAsset(
         .doc(currentStockName);
     return FirebaseFirestore.instance.runTransaction((transaction) async {
       DocumentSnapshot snapshot = await transaction.get(documentReference);
-      var newVolume = double.parse(snapshot['Volume']) - value;
-      var newCost = double.parse(snapshot['Cost']) - cost;
-      var newPrice = newCost / newVolume;
-      if (newVolume == 0) {
-        documentReference.delete();
-        response = ('You have sold all Assets in $currentStockName');
-        return (response);
-      } else if (newVolume <= 0) {
-        response = ("you don't have enough $currentStockName assets");
-        return (response);
+      if (snapshot.exists) {
+        var newVolume = double.parse(snapshot['Volume']) - value;
+        var newCost = double.parse(snapshot['Cost']) - cost;
+        var newPrice = newCost / newVolume;
+        if (newVolume == 0) {
+          documentReference.delete();
+          response = ('You have sold all Assets in $currentStockName');
+          return (response);
+        } else if (newVolume <= 0) {
+          response = ("you don't have enough $currentStockName assets");
+          return response;
+        }
+        transaction.update(documentReference, {
+          'Volume': newVolume.toString(),
+          'Price': newPrice.toString(),
+          'Cost': newCost.toString()
+        });
+        response = "Success";
+        return response;
+      } else {
+        return "Error! You Have No $currentStockName Asse";
       }
-      transaction.update(documentReference, {
-        'Volume': newVolume.toString(),
-        'Price': newPrice.toString(),
-        'Cost': newCost.toString()
-      });
-      response = "Success";
-      return response;
     });
   } catch (e) {
     return 'failed';
